+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/metadata/MetadataImportExportController.java
@@ -81,9 +81,9 @@ import org.hisp.dhis.user.CurrentUserService;
 import org.hisp.dhis.user.CurrentUserUtil;
 import org.hisp.dhis.user.UserSettingKey;
 import org.hisp.dhis.user.UserSettingService;
-import org.hisp.dhis.util.FileUtils;
 import org.hisp.dhis.webapi.mvc.annotation.ApiVersion;
 import org.hisp.dhis.webapi.service.ContextService;
+import org.hisp.dhis.webapi.utils.FileResourceUtils;
 import org.springframework.beans.factory.ObjectFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -157,7 +157,7 @@ public class MetadataImportExportController
         throws IOException,
         ConflictException
     {
-        File tempFile = FileUtils.toTempFile( request.getInputStream() );
+        File tempFile = FileResourceUtils.toTempFile( request.getInputStream() );
 
         MetadataImportParams params = metadataImportService.getParamsFromMap( contextService.getParameterValuesMap() );
